// src/types/film.ts
export type Film = {
  id: string;
  name: string;
  genre: string;
  posterImage: string;
  backgroundImage: string;
  videoLink: string;
  released: number;
  isFavorite: boolean;
  backgroundColor: string;
  description: string;
  rating: number;
  scoresCount: number;
  director: string;
  starring: string[];
  runTime: number;
  previewImage: string;
  previewVideoLink: string;
};